LiveTalking/avatar/requirements.txt
livekit==1.0.7
livekit-agents[openai,silero,deepgram,elevenlabs]==1.0.20
requests
python-dotenv
pandas

